% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_sim.R
\name{collapse_and_combined}
\alias{collapse_and_combined}
\title{Function to resize patches in a SpatRaster using the terra and landscapemetrics packages}
\usage{
collapse_and_combined(patched_raster, og_raster)
}
\arguments{
\item{patched_raster}{The SpatRaster object representing the categorical data.}

\item{og_raster}{The values of the classes of interest, default is "all".}
}
\value{
A SpatRaster where each patch is defined as a unique value and each
class has its own layer in the spatraster.
}
\description{
Function to resize patches in a SpatRaster using the terra and landscapemetrics packages
}
