% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_conf_mat.R
\name{download_cdl_mat_files}
\alias{download_cdl_mat_files}
\title{Function to Read in and upzip confusion matrix files from the web}
\usage{
download_cdl_mat_files(years, temp_dir = "extracted_files")
}
\arguments{
\item{years}{The years of data that you want to be downloaded.}

\item{temp_dir}{The file name you want the extracted files to be stored in.}
}
\value{
The raw excel books containing the confusion matrix data for
every US state's USDA Crop Land Data Layer.
}
\description{
Function to Read in and upzip confusion matrix files from the web
}
