% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_sim.R
\name{find_patches}
\alias{find_patches}
\title{Function to resize patches in a SpatRaster using the terra and landscapemetrics packages}
\usage{
find_patches(raster)
}
\arguments{
\item{raster}{The SpatRaster object representing the categorical data.}
}
\value{
A SpatRaster where each patch is defined as a unique value and each
class has its own layer in the spatraster.
}
\description{
Function to resize patches in a SpatRaster using the terra and landscapemetrics packages
}
