% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_sim.R
\name{generate_transition_vectors}
\alias{generate_transition_vectors}
\title{Function to generate transition vectors for negative cell values}
\usage{
generate_transition_vectors(r, transition_matrix, iterations = 10)
}
\arguments{
\item{r}{The SpatRaster object from get_patches with one layer for each
class.}

\item{transition_matrix}{The transition matrix you want to use.}
}
\value{
A matrix of vector for each patch in each layer.
}
\description{
Function to generate transition vectors for negative cell values
}
