% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_conf_mat.R
\name{get_mat_data}
\alias{get_mat_data}
\title{Function to confusion matrix data for your state of interest}
\usage{
get_mat_data(state_abbreviation)
}
\arguments{
\item{state_abbreviation}{A two letter abbreviation for US states.}
}
\value{
The data frames representing the confusion matrices for your states
of interest for all years previously downloaded.
}
\description{
Function to confusion matrix data for your state of interest
}
