% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_conf_mat.R
\name{get_mat_data_dep}
\alias{get_mat_data_dep}
\title{Function to retrieve confusion matrix data for multiple states of interest}
\usage{
get_mat_data_dep(
  state_abbreviation,
  file_path = "inst/extdata/extracted_files",
  verbose = FALSE
)
}
\arguments{
\item{state_abbreviation}{A vector of two-letter abbreviations for US states.}

\item{file_path}{The path to the directory where files are stored
(default is "inst/extdata/extracted_files").}

\item{verbose}{The stops the messages from printing to the console.}
}
\value{
A named list where each element is a list of data frames representing
confusion matrices for each state.
}
\description{
Function to retrieve confusion matrix data for multiple states of interest
}
