% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_conf_mat.R
\name{get_trans_mat}
\alias{get_trans_mat}
\title{Function to format the confusion matrices as transition matrices}
\usage{
get_trans_mat(df_list, categories)
}
\arguments{
\item{df_list}{A list of lists of data frames extracted using get_mat_data(). Each sublist should contain two or more data frames to be summed.}

\item{categories}{A list of categories defining the numbers between 1 and 256.}
}
\value{
A list of data frames where row names represent pixels that will
transition and column names represent the class they will transition to.
}
\description{
Function to format the confusion matrices as transition matrices
}
