% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_sim.R
\name{modify_matrix}
\alias{modify_matrix}
\title{Function to modify transition matrix based on user input.}
\usage{
modify_matrix(mat, indices = 0)
}
\arguments{
\item{mat}{The transition matrix the you want to update.}

\item{indices}{The vector of classes you want to remain unsimulated.}
}
\value{
A transition matrix modified to not transition some classes.
}
\description{
Function to modify transition matrix based on user input.
}
