% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_rasters.R
\name{process_single_layer}
\alias{process_single_layer}
\title{Function to help process SpatRasters by layer}
\usage{
process_single_layer(layer, confusion_matrix)
}
\arguments{
\item{layer}{The pixel values from each layer of the tagged SpatRaster.}

\item{confusion_matrix}{A confusion matrix created from get_trans_mat() that
represents the transition probabilities.}
}
\value{
New pixel values for the edge pixels.
}
\description{
Function to help process SpatRasters by layer
}
