% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_rasters.R
\name{simulate_raster_changes}
\alias{simulate_raster_changes}
\title{Function to simulate raster pixels on edges}
\usage{
simulate_raster_changes(
  input_raster,
  edge_depth = 1,
  n_simulations = 10,
  confusion_matrix
)
}
\arguments{
\item{input_raster}{Original reclassified raster.}

\item{edge_depth}{The number of cells away from the patch edge to be
considered a core cell.}

\item{n_simulations}{The number of simulation you want to run.}

\item{confusion_matrix}{A confusion matrix created from get_trans_mat() that
represents the transition probabilities.}
}
\value{
The simulated CDL SpatRaster objects.
}
\description{
Function to simulate raster pixels on edges
}
