% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_sim.R
\name{simulate_raster_patch}
\alias{simulate_raster_patch}
\title{Function to simulate tagged patch values in a single layer spatraster.}
\usage{
simulate_raster_patch(original_raster, transition_matrix, iterations = 10)
}
\arguments{
\item{original_raster}{The SpatRaster object representing the categorical data.}

\item{transition_matrix}{Transition matrix values that define transitions based on class values.}

\item{iterations}{The number of simulations to be performed.}
}
\value{
A SpatRaster where the negative value patches are set to transition.
}
\description{
Function to simulate tagged patch values in a single layer spatraster.
}
