% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_rasters.R
\name{tag_and_transition}
\alias{tag_and_transition}
\title{Function to tag edges and apply transitions.}
\usage{
tag_and_transition(input_raster, edge_depth, confusion_matrix)
}
\arguments{
\item{input_raster}{Original reclassified raster.}

\item{edge_depth}{The number of cells away from the patch edge to be
considered a core cell.}

\item{confusion_matrix}{A confusion matrix created from get_trans_mat() that
represents the transition probabilities.}
}
\description{
Function to tag edges and apply transitions.
}
