% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_rasters.R
\name{tag_edges}
\alias{tag_edges}
\title{Function to identify and tag edges using the terra package}
\usage{
tag_edges(raster, edge_depth = 1)
}
\arguments{
\item{raster}{The SpatRaster object representing the USDA's Crop Land Data
Layer.}

\item{edge_depth}{The number of cells away from the patch edge to be
considered a core cell.}
}
\value{
A raster where negative numbers represent the edge pixels.
}
\description{
Function to identify and tag edges using the terra package
}
