% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_terra_sim.R
\name{tag_edges_2layer}
\alias{tag_edges_2layer}
\title{Function to identify and tag edges using the terra package}
\usage{
tag_edges_2layer(raster, edge_depth = 1)
}
\arguments{
\item{raster}{The SpatRaster object representing the USDA's Cropland Data
Layer.}

\item{edge_depth}{The number of cells away from the patch edge to be
considered a core cell.}
}
\value{
A  2 layers raster where negative numbers represent the edge pixels.
}
\description{
Function to identify and tag edges using the terra package
}
