% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_rasters.R
\name{transition_pixels}
\alias{transition_pixels}
\title{Function to transition values of "edge" pixels based on confusion matrices}
\usage{
transition_pixels(pixel_values, confusion_matrix)
}
\arguments{
\item{pixel_values}{The pixel values from the tagged SpatRaster.}

\item{confusion_matrix}{A confusion matrix created from get_trans_mat() that
represents the transition probabilities.}
}
\value{
New pixel values for the edge pixels.
}
\description{
Function to transition values of "edge" pixels based on confusion matrices
}
